% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{attach_attribs}
\alias{attach_attribs}
\title{Attach 'router-hidden' class to single page UI content}
\usage{
attach_attribs(ui, path)
}
\arguments{
\item{ui}{Single page UI content created with proper html tags or tag list.}

\item{path}{Single page path name. Attached to \code{data-path} attriubute.}
}
\description{
Covered UI types are Shiny/htmltools tags or tag lists and httml templates.
In case of tag list (\code{tagList}) and html template (\code{htmlTemplate}) 'div' wrapper
with 'router-hidden' class is added.
}
