% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{callback_mapping}
\alias{callback_mapping}
\title{Create a mapping bewtween a ui element
and a server callack}
\usage{
callback_mapping(ui, server = NA)
}
\arguments{
\item{ui}{Valid Shiny user interface.}

\item{server}{Function that is called within the global server function if given}
}
\description{
Create a mapping bewtween a ui element
and a server callack
}
