% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{create_router_callback}
\alias{create_router_callback}
\title{Internal function creating a router callback function.
One need to call router callback with Shiny input and output in server code.}
\usage{
create_router_callback(root, routes)
}
\arguments{
\item{root}{Main route to which all invalid routes should redirect.}

\item{routes}{A routes (list).}
}
\value{
Router callback.
}
\description{
Internal function creating a router callback function.
One need to call router callback with Shiny input and output in server code.
}
