% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{disable_bootstrap_on_bookmark}
\alias{disable_bootstrap_on_bookmark}
\title{Fix conflicts when some bookmark uses bootstrap}
\usage{
disable_bootstrap_on_bookmark(bookmark)
}
\arguments{
\item{bookmark}{Bookmark name on which bootstrap dependency should be suppressed.}
}
\description{
This function dynamically removes bootstrap dependency when user opens specified bookmark.
It should be inserted in head of bootrstrap page.
}
