% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{On Load}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{library name}

\item{pkgname}{package name}
}
\description{
On package load it updates .i18_config reading yaml file from config.
}
\keyword{internal}
