% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{extract_link_name}
\alias{extract_link_name}
\title{Extract link name}
\usage{
extract_link_name(path)
}
\arguments{
\item{path}{character with link path}
}
\value{
stripped link
}
\description{
Strips off the first 3 character, assuming that they are: "#!/".
}
