% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_query_param}
\alias{get_query_param}
\title{Get Query Parameters}
\usage{
get_query_param(field = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{field}{If provided, retrieve only a param with this name. (Otherwise,
return all params)}

\item{session}{The Shiny session}
}
\value{
The full list of params on the URL (if any), as a list. Or, the single
requested param (if present). Or NULL if there's no input, or no params.
}
\description{
Convenience function to retrieve any params that were part of the requested
page. The param values returned come from "httr::parse_url()"
}
