% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{get_url_hash}
\alias{get_url_hash}
\title{Internal function to get url hash with #!.}
\usage{
get_url_hash(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{The current Shiny Session}
}
\value{
Reactive hash value.
}
\description{
Internal function to get url hash with #!.
}
