% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{\%:::\%}
\alias{\%:::\%}
\title{::: hack solution}
\usage{
pkg \%:::\% name
}
\arguments{
\item{pkg}{package name}

\item{name}{function name}
}
\value{
function
}
\description{
::: hack solution
}
\keyword{internal}
