% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{is_page}
\alias{is_page}
\title{Is page}
\usage{
is_page(page, session = shiny::getDefaultReactiveDomain(), ...)
}
\arguments{
\item{page}{The page to display. Should match one of the paths sent to the}

\item{session}{Shiny session}

\item{...}{Other parameters are sent through to shiny::req()
router.}
}
\description{
Tell the reactive chain to halt if we're not on the specified page. Useful
for making sure we don't waste cycles re-rendering the UI for pages that are
not currently displayed.
}
