% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_msg}
\alias{log_msg}
\title{Helper function to print out log messages into Shiny
using cat() and stderr(), as described on
https://shiny.rstudio.com/articles/debugging.html}
\usage{
log_msg(...)
}
\arguments{
\item{...}{All params get passed through to cat().
They're automatically wrapped in shiny::isolate(),
so you can print reactive values here without too
much worry.}
}
\description{
Because this can print a lot, it's silent unless
the shiny.router.debug option is set.
}
