% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{make_router}
\alias{make_router}
\title{Creates router.}
\usage{
make_router(default, ...)
}
\value{
Shiny router callback that should be run in server code with Shiny input and output lists.
}
\description{
Creates router.
}

