% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{page404}
\alias{page404}
\title{404 page}
\usage{
page404(page = NULL, message404 = NULL)
}
\arguments{
\item{page}{shiny page style, eg. \code{shiny::tags$div(h1("Not found"))}}

\item{message404}{message to display at the 404 website}
}
\description{
The page which appear when path is wrong.
}
\examples{
page404() # shiny::tags$div(h1("Not found"))
page404(message404 = "ABC") # shiny::tags$div(h1("ABC"))
}
