% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\docType{data}
\name{PAGE_404_ROUTE}
\alias{PAGE_404_ROUTE}
\title{Default 404 page}
\format{An object of class \code{character} of length 1.}
\usage{
PAGE_404_ROUTE
}
\description{
This is default 404 page.
}
\keyword{datasets}
