% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{parse_url_path}
\alias{parse_url_path}
\title{Parse url and build GET parameters list}
\usage{
parse_url_path(url_path)
}
\arguments{
\item{url_path}{character with link url}
}
\value{
list containing two objects: \itemize{
 \item path
 \item query, a list
}
}
\description{
Extract info about url path and parameters that follow \code{?} sign.
}
\details{
\code{parse_url_path} allows parsing paramaters lists from url. See more in examples.
}
\examples{
parse_url_path("#!/?a=1&b=foo")
parse_url_path("#!/?a=1&b[1]=foo&b[2]=bar")
parse_url_path("www.foo.bar?a=1&b[1]=foo&b[2]=bar")
}
