% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{route}
\alias{route}
\title{Create single route configuration.}
\usage{
route(path, ui, server = NA)
}
\arguments{
\item{path}{Website route.}

\item{ui}{Valid Shiny user interface.}

\item{server}{Function that is called as callback on server side}
}
\value{
A route configuration.
}
\description{
Create single route configuration.
}
\examples{
route("/", shiny::tags$div(shiny::tags$span("Hello world")))

route("/main", div(h1("Main page"), p("Lorem ipsum.")))
}
