% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{router_server}
\alias{router_server}
\title{Create router pages server callback}
\usage{
router_server(router)
}
\arguments{
\item{router}{Router pages object. See \link{make_router}.}
}
\description{
Create router pages server callback
}
\keyword{internal}
