% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{router_ui_internal}
\alias{router_ui_internal}
\title{Creates router UI}
\usage{
router_ui_internal(router)
}
\arguments{
\item{router}{Router pages object. See \link{make_router}.}
}
\value{
list with shiny tags that adds "router-page-wrapper" div and embeds
router javascript script.
}
\description{
Creates router UI
}
\keyword{internal}
