% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{valid_path}
\alias{valid_path}
\title{Internal function that validates that path is defined in routes.}
\usage{
valid_path(routes, path)
}
\arguments{
\item{routes}{A routes (list).}

\item{path}{A path.}
}
\value{
Boolean value indicating if path is defined.
}
\description{
Internal function that validates that path is defined in routes.
}
\keyword{internal}
