% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{cleanup_hashpath}
\alias{cleanup_hashpath}
\title{Formats a URL fragment into a hashpath starting with "#!/"}
\usage{
cleanup_hashpath(hashpath)
}
\arguments{
\item{hashpath}{character with hash path}
}
\description{
Formats a URL fragment into a hashpath starting with "#!/"
}
