% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{escape_path}
\alias{escape_path}
\title{Escapes routing path from not safe characters.}
\usage{
escape_path(path)
}
\arguments{
\item{path}{A path.}
}
\value{
String with escaped characters.
}
\description{
Escapes routing path from not safe characters.
}

