% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{route_link}
\alias{route_link}
\title{Route link}
\usage{
route_link(path)
}
\arguments{
\item{path}{character with path}
}
\value{
route link
}
\description{
Adds /#!/ prefix to link.
}
\examples{
route_link("abc") # /#!/abc
}
