% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{router_server}
\alias{router_server}
\title{Create router pages server callback}
\usage{
router_server(root_page = "/", env = parent.frame())
}
\arguments{
\item{root_page}{Main page path.}

\item{env}{Environment (only for advanced usage).}
}
\value{
Router pages server callback.
}
\description{
Server part of the router.
}
\examples{
\dontrun{
  server <- function(input, output, session) {
    router_server(root_page = "/")
  }
}
}
