% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{valid_path}
\alias{valid_path}
\title{Validates that path is defined in routes.}
\usage{
valid_path(routes, path)
}
\arguments{
\item{routes}{A routes (list).}

\item{path}{A path.}
}
\value{
Boolean value indicating if path is defined.
}
\description{
Validates that path is defined in routes.
}

